Bugfix: Invalidate cache after `num_hypotheses` changes
After the `num_hypotheses` changes, we possibly have cached translations that
are searched with different beam_search parameters, and also stored in
the lists with different lengths.
The latter becomes problem when combining paragraphs.
So in this case we re-do the search with new parameter(s), and cache the result.